chore: split out mock stub
