+++ packages/@lwc/wds-core/src/node/plugins/mock-esm/util.js
@@ -1,0 +1,2 @@
+export const withoutDefault = (strings) => strings.filter((el) => el !== 'default');
+export const hasDefault = (strings) => strings.includes('default');
